Тупко #29: Windows callback - fourth step
