+++ openssl/pubnub_config.h
@@ -202,7 +202,6 @@
 #define PUBNUB_USE_ACTIONS_API 1
 #endif
 
-#if !defined _WIN32
 #if !defined(PUBNUB_USE_AUTO_HEARTBEAT)
 /** If true (!=0) will enable using the Auto Heartbeat Thumps(beats), which is a feature
     that enables keeping presence of the given uuids on channels and channel groups during
@@ -211,6 +210,5 @@
     these things all by himself using pubnub_heartbeat() transaction */
 #define PUBNUB_USE_AUTO_HEARTBEAT 1
 #endif
-#endif /* !defined _WIN32 */
 
 #endif /* !defined INC_PUBNUB_CONFIG */
