+++ windows/pbtimespec_elapsed_ms.h
@@ -2,7 +2,6 @@
 #if !defined(INC_PBMC_ELAPSED_MS)
 #define INC_PBMC_ELAPSED_MS
 
-#include <winsock2.h>
 #include <windows.h>
 
 int pbtimespec_elapsed_ms(FILETIME prev_timspec, FILETIME timspec);
